requirements.txt
fastapi
uvicorn
langchain
langchain-community
langgraph
chromadb
sentence-transformers
google-generativeai
streamlit
requests
python-dotenv